testdrive: more aggressive retry of SQL statements
In the typical CREATE -> inject -> SELECT pattern, it takes
several seconds for the SELECT to return the right result because
the retry strategy is not aggressive enough and quickly begins to
perform multi-second sleeps.

Modify retry.rs to allow for a custom exponential factor to be
specified so that testdrive can use factor 1.5 and a starting backoff
value of 50ms.

This change alone shaves a second or more from each eligible SELECT.