Support `ENVELOPE DEBEZIUM UPSERT`
This adds upsert semantics support for Debezium sources. Fundamentally that means that we treat
upstream diffs as potentially nonsensical.

To do that, we track keys per-key values per-key, and only retract rows that we know that we have
seen.

There is one todo in the code that will be the immediate next work: with the code as it stands we
treat the keys as raw bytes, not decoding them, but since Avro is not self-describing it is possible
for the same bit pattern to have different values when decoded. We feel like this is unlikely enough
to be an issue for most users (considering especially that we don't really support schema evolution)
that it's probably safe to use as is.