kafka: wire exactly_once setting to sink code, use it to set transactional.id
This is the first time our exactly-once sink actually uses transactions.
The fixes in the previous commits are required to make this work.

Before, the sink code was assuming that a `transactional.id` was set on
the Kafka config, but this was never the case.

Now, we set a `transactional.id` if exactly_once=true.

This only works as long as there is only one real Kafka Producer
instance that is producing data. If/when we allow multiple parallel
producers we need to add a suffix to differentiate them.