materialized: fix panic with closed HTTP connections
An HTTP connection that closes at the wrong time can cause Materialize
to panic. The cause is that Hyper was drops HTTP connection handler
futures when the client goes away, in an effort to avoid doing
unnecessary work. But we have important cleanup work that must be
performed when a connection closes; failing to do that work is in fact
panicworthy.

Unfortunately Hyper does not allow you to request that all futures be
polled to completion, even if the client goes away.  So this commit
fixes the bug by introducing a new `spawn_if_canceled` future
combinator. The idea is that we can use the `Drop` implementation of
this combinator to call `tokio::spawn` to poll the original future to
completion. It's a bit weird, but it seems to work well.

An alternative solution is to use an `Arc<Mutex<SessionClient>>` to
share the SessionClient outside of the future, so that code outside the
future could perform the necessary cleanup. That turns out to be much
messier, though, and has to be done for every type inside the future
that requires this sort of cleanup, which includes a channel deep inside
the "/sql" endpoint handler.

The "correct" solution to this problem would use asynchronous
destructors (i.e., `AsyncDrop`), but that's not presently a thing in
Rust, and probably won't be for years.

Huge thank you to @ruchirK and @antifuchs for helping to track this
down.

Fix #6278.