Update Avro schema parsing to be identical everywhere
This supports using Schema Registry URLs, Key Schemas and inline/file schemas in the two places that
they can be used. This does *not* change the syntax of `ENVELOPE UPSERT`, which still allows
specifying an extra `FORMAT AVRO` after itself, a later PR will deprecate that and use the new
unified syntax.

This allows optionally specifying Avro inline value schemas as `VALUE SCHEMA` instead of just
`SCHEMA` or requiring `VALUE SCHEMA`, to merge the difference between the CSR syntax and the
inline-only syntax.

This should correctly populate the key schema from the CSR if the schema exists, I'll test that in
the follow-up PR where we start actually using this data.

This doesn't change any semantics of any other part of the system, and detailed testing will come
with the follow-up PR, things are left as-is for now to ensure that this doesn't change any
behavior.