Use named volumes unless MZ_CHBENCH_SNAPSHOT is set
We've been having issues with the `./tmp` directory causing startup
issues with Zookeeper, failures in CI and, by default, the directory is
created with root permissions. By using the ":+" interpolation, we can
use named volumes in all situations where MZ_CHBENCH_SNAPSHOT is unset.

When MZ_CHBENCH_SNAPSHOT is set, we can tack on the subdirectory using
":+" interpolation, preserving the existing behavior.