"fix" upsert for the new "decoders know how to decode themselves" world
Upsert does a bunch of work to put the key data into the value data, which is the next thing on my
list to get rid of. This change works around the fact that the `DecoderState::decode_key` method
used to be basically an alias for `decode_upsert_value`, but with the previous commit it actually
uses a different decoder.

In-progress work will clean this up.