coord,sql: rework index nullability tracking
Store index nullability information in catalog's `Index` struct
directly, rather than trying to compute it on the fly. This greatly
simplifies the system table reporting flow.

Also, rename the index "keys" field to "fields". A "key" is a synonym
for "index"; using "key" to mean "column within an index" is quite
confusing. The new terminology is in line with PostgreSQL.