Record a prometheus metric counting the log messages emitted by level
This change adds a new metric, "mz_log_message_total", which counts the
number of log messages emitted by level. This should make it easier for
users of the cloud product (and on-prem users also) to hook up alerting
to the log messages: Our recommendation is to page people if an ERROR is
logged, so this allows doing just that.