Limit magit commit message buffers to 72 chars in .editorconfig
This is a bit of a hack, but annoying otherwise: Our default line length
limit is 100 chars, so anytime you make a commit with magit, it would
word-wrap at 100 chars. By setting a rule for COMMIT_EDITMSG, this
causes the commit message editor to auto-wrap at the right spot again.