Drop topics before starting debezium connector (#6421)
Chbench startup would often timeout because Peeker failed to create
sources due to the Kafka topics missing. This is because we drop the
Kafka topics after creating them in the connector. This then results in
Peeker either succeeding or failing depending on how quickly Debezium
would recreate the topics.

Instead of dropping the topics after starting the connect process,
simply drop them beforehand. That way the topics will be empty and will
exist for both debezium and peeker.