render: optimize topk when limit=1 and input is monotonic (append-only)
If the input to a TopK is monotonic (aka append-only aka no retractions)
then we don't have to worry about keeping every row we've seen around
forever. Instead, the reduce can incrementally compute a new answer by
looking at just the old TopK for a key and the incremental data. We do
an identical optimization for MIN and MAX of monotonic sources.

This optimization generalizes to any TopK over a monotonic source, but
we special case only TopK with offset=0 and limit=1 (aka Top1) for now.
This is because (1) Top1 can merge in each incremental row in constant
space and time while the generalized solution needs something like a
priority queue and (2) we expect Top1 will be a common pattern used to
turn a Kafka source's "upsert" semantics into differential's semantics.
(2) is especially interesting because Kafka is monotonic with an
ENVELOPE of NONE, which is the default for ENVELOPE in Materialize and
commonly used by users.

Closes #5996