coord: use AntichainTokens to propagate frontier since changes
This is a refactoring to allow for a future feature, but this commit
does not affect any user-facing behavior.

Previously, update_upper would immediately queue compaction messages
in since_updates. Upcoming read transactions need the ability to hold
back compactions at some timestamp until the transaction is completed.
Change Frontiers' since to be modifiable with its since_handles
function. That function wraps a MutableAntichain that can be advanced
and forwards its changes to the Frontiers' since, including (and most
importantly) forwarding removals when Dropped.

Establish a single since_handle per index/source that can be advanced
by update_upper. In the future, any number of other things (like read
transactions) can also get their own since_handle on an index that
will be dropped when the need to prevent compaction of the timestamp
is finished.