Deduplicate identical ports on docker inspect output
After a recent upgrade of Docker, I got the following error when trying
to start an mzcompose workflow:

    Failed! Could not unambiguously determine port for mysql found: 3306,3306

Looking at the output of `docker inspect`, we now see this:

    "Ports": {
      "3306/tcp": [
        {
          "HostIp": "0.0.0.0",
          "HostPort": "3306"
        },
        {
          "HostIp": "::",
          "HostPort": "3306"
        }
      ],
      "33060/tcp": null
    },

It appears that Docker is listing ipv4 and ipv6 addresses as their own
entries. For the purposes of find_host_ports, which uses `localhost`,
let's treat these identically.