kafka: use natural (primary) relation key for sinks if available
This changes how `before` and `after` updates for DEBEZIUM sinks are
merged together. We now do the following:

1. if there is a user-specified key, merge updates by that (as before),
   otherwise
2. if there is a natural (primary) relation key, merge updates by that,
   otherwise
3. consolidate using whole row, and don't merge updates.

For cases 2. and 3. we don't write the key out to Kafka but only use it
to consolidate/merge updates and distribute the sink work across
workers.

Using a key for all cases has the effect of more evenly utilizing all
available workers, increasing throughput.

We need to adapt some tests because the order of writing unrelated
changes to Kafka is not deterministic anymore. These tests were working
before because both the sources and the sinks where non-parallel, which
is not necessarily the case for real-world use cases. Now the tests
behave more like those real-world use cases. Otherwise, user-visible
behaviour should not change, other than the aforementioned improved
throughput out of the box.

This also adds test for all the possible combinations of relation key
and user-specified key for DEBEZIUM sinks. Only DEBEZIUM sinks are
effected because 1) UPSERT sinks always have a user-specified key, and
2) this only affects Kafka sinks for now.