sql: integrate postgres source namespace into table string
Previously we used a NAMESPACE string to specify the namespace/schema
of a postgres replication source. For most sources this was a
redundant "public" specifier that no one cared about. It was needed to
be able to determine the OID of the table. We can instead parse the
table as an object name, and then convert it to an ast-safe string
that can include a namespace if needed by a user.

This greatly simplifies the syntax and should feel more native to
users of materialize, who might otherwise be surprisied that in this
one specific place, you used to have to quote your table name.