tests: use 5 workers for testdrive tests
We do this to provoke tests that would only fail with higher numbers of
workers. Specifically, the tests in avro-sinks.td verify the consistency
topic but we recently fixed a bug that only occurs when there is more
than one worker.

Increasing the number of workers is only possible with the previous
commits that ensure only one worker is writing to Kafka.