source/postgres: present the raw replication stream
Previously the postgres source was both reading the replication stream
and decoding the rows of a particular table to present them to the
system. This makes replicating multiple tables efficiently tricky.

This change makes the source present the raw replication stream to the
system for futher processing. The output schema is (oid, row_data) where
oid is the relation id and row_data is a LIST of the column data encoded
as text. Casting the text data in the appropriate type is the
responsibility of downstream consumers.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>