Allow error values from parse_debezium
Instead of print error messages to the log (or, perhaps even worse,
panic'ing), we should indicate, via error types, that we failed to
parse the transaction message.

This doesn't change the caller of parse_debezium and, if we like where
this is going, I can start pulling on that thread.

We probably also want to convert these error types to structured error
types.