sql: teach functions about volatility and fixup OIDs
Add volatility to all functions. This is currently unused but will be
in future commit.

Add an ignored test that can be used to verify correctness of any
postgres-identical functions we have. This test uncovered a handful of
small OID and other errors which have also been corrected here.

This test enforces that we match postgres volatility completely. However
postgres marks many functions as stable because their output depends on
session settings (timezone, for example). I think today we don't allow
changing any of the ones that could affect output, so the functions
for us are currently immutable. In this commit, since volatility isn't
used yet, we can "downgrade" our volatility to be postgres compatible,
which will make any changes we decide on more explicit.