include the Kafka topic name when calling fetch_metadata in purification
During purification of Kafka CREATE SOURCE we make a fetch_metadata call
to validate basic connectivity/auth, allowing us to fast-fail to the
user with a nice error message. This change scopes that fetch_metadata
call to the specific topic we're creating a source from in an effort to
play nicer with whatever Kafka ACLs may be configured.

Note that this is not appropriate for eg. sinks where the topic may not
have been created. Calling fetch_metadata with a specific nonexistent
topic name may _create_ that topic in certain Kafka cluster
configurations.