coord,sql: auto-generate an available name for default indexes
Previously, when creating the default index for a materialized source,
view, or table, Materialize would insist upon using the name
"OBJECT_primary_idx". If this name was not available, the entire
operation was aborted.

This was at odds with the behavior of "CREATE INDEX", which will search
until it finds an available name. If "OBJECT_primary_idx" is not
available, Materialize will try "OBJECT_primary_idx1",
"OBJECT_primary_idx2", and so on.

This commit teaches "CREATE MATERIALIZED {SOURCE,VIEW}" and "CREATE
TABLE" to do the same. In particular, this makes the following sequence
of statements succeed where it used to fail:

    CREATE TABLE t_orig ()
    ALTER TABLE t_orig RENAME TO t_dontcare
    CREATE TABLE t_orig ()

Fix #5096, though not exactly as proposed in the issue. (The fix
described there suggested renaming indexes when the object they're built
on is renamed, but that would be at odds with PostgreSQL and is also
just hard.)