mzcompose: accept --mz* arguments anywhere
Make it possible to say things like:

    ./mzcompose run service --mz-build-mode=dev

Previously this would attempt to pass "--mz-build-mode=dev" as an argument
to the service; you had to instead put the argument first, as in

    ./mzcompose --mz-build-mode=dev run service

which was irritating.

The solution was to change the nargs specifier from "argparse.REMAINDER"
to "*", as the former inhibits parsing of known options.