doc: design for managing external state in Materialize
As we support sources and sinks that will create, track, and destroy
state external to Materialize, it would be best to interact with that
state in a uniform way. This design doc puts forward a proposal that
was discussed and agreed upon by the External Integrations team. No
concrete code implementing this idea has been merged yet.