Relax restriction on zero-collection transactions
I don't know why our test cases have an example transaction with
zero collections. Instead of changing the test case, let's relax
the error condition in Materialize until I understand that better.

Instead, replace the check with a version that verifies that the
sum of the collection events matches the event_count total.