sql,sql-parser: remove CREATE SOURCES FROM POSTGRES
As per our discussions last week, the design of Postgres sources has
shifted. Instead of supporting CREATE SOURCES FROM POSTGRES, we are
going to require users to create a single source from an upstream
publication. Once that source has been created, users will be able to
run a (yet to be created) CREATE MATERIALIZED VIEWS statement on top
of their Postgres source. This is the first step in this change.