kafka-util: no longer canonicalize schemas in kgen
The Avro spec provides conflicting/unclear guidance on
canonicalizing schemas that contain decimal data (i.e. decimal
values require the presence of precision and scale fields to
express equality, but neither field is documented in the
canonicalization process).

To solve this problem for kgen, we are no longer canonicalizing
schemas.