kafka: verify that (user-specified) UPSERT sink key is valid for the input collection
This changes our Kafka sink to only allow keys that are unique keys for
the input collection.

Without this property, it can happen that there are multiple updates at
a given timestamp that match the upsert key. We cannot currently handle
this because we cannot decide which update should take precedence.