Remove duplicate Grafana dashboards (#6683)
* Remove duplicated Grafana dashboard

This dashboard has been copy/pasted around to a number of benchmarks
but it superceded by the default materialized dashboard already present
in the monitoring container. Let's just remove it.

It's useful in chbench because it shows metrics gathered from peeker and
perhaps we should rename it `peeker.json` to reflect that?

* Remove persistent volume for grafana

They don't need to be persistent, as the dashboard definitions are
already predefined in the monitoring container. Any updates to these
dashbaords should go there.