test: change consolidation.td to not race against update_upper
Previously, we would create a materialized source, insert some data and then
create a sink afterwards. This causes problems because by that time, the data
sometimes had already been ingested, and the requisite upper is already updated,
which means the sink won't get the updates at earlier times as the test requires.

The fix here is to change sink creation to occur before we send any data, guaranteeing
that it will be made as of 0.