pgwire: correctly clear txn state during errors
Previously the call to clear_transaction ignored the sinks that needed
to be dropped. Prevent this by adding a must_use directive so we don't
make this mistake again in the future. Now the error route correctly
goes through the coordinator which can process any needed side effects
when ending a transaction.