Resurrect chbench consistency checks
This commit makes a number of changes aimed at bring back the
correctness checker for chbench / debezium replication / consistent
sources and running it nightly under CI.

The correctness checker fails for both source database types: MySQL and
Postgres. Add this to the nightly CI job but mark the tests as soft_fail
to prevent them from failing the overall build. Let's hope that we can
start making these pass once #6296 is fixed!

When running with MySQL, we're currently getting an error parsing the
consistency topic that causes the timestamper thread to exit. When
running with Postgres, no Materialized errors are encountered but the
queries fail to return correct results.

Default to running all correctness checks and remove some stale
configuration / README files.

Fail the correctness checker program if any of the checks fail and
assert that the container exits cleanly within the workflow.

Bump the Debezium example images to 1.5 because 1.5 includes the changes
necessary to enable the transaction (consistency) topic for MySQL.