dataflow: respect `Context::as_of_frontier` in `SourceConnector::Local`
The `as_of_frontier` should be applied to sources to improve performance
and prevent incorrect results in sinks. So far, it has not been an
actual issue for `SourceConnector::Local` as users are not able to
provide timestamps for inserted records. Future changes might have led
to surprises though.

Fixes #5865.