sinks: relax UPSERT key checking, improve error message and warning
In #6709 we introduced strict checks for the specified upsert sink key.
It turns out those checks are too strict because the optimizer currently
cannot (or will not) figure out the key for some views/collections where
it is clear for a user to see that there is a valid upsert key.

Side note, a valid upsert key must be a primary key of the sinked
collection.