coord: prevent views from multiple timelines
Teach the system about Timelines, which is the meaning of the timestamp
number. Prevent users from joining data from multiple timelines because
it's meaningless and/or will block forever.

The implementation here avoids teaching ship_dataflow or DataflowBuilder
about timelines because those require transactional catalog changes which
we don't yet have, and are not straightforward to implement. Instead
we opt-in the places that can create arbitrary SQL expressions (SELECT,
CREATE VIEW) to check for timeline violations.

These changes introduce the coordinator to the notion of timelines and block queries
would span multiple timelines.