persist: begin initial skeleton of persistence core subsystem
Merging early to faciliate collaboration.

A persistence user would start by constructing a `PersistManager` which
is a thin `Send+Sync+Clone` handle around a `Persister` impl which does
the heavy lifting (see details below). `PersistManager` is a registry
for multiple `Persisted`s (each corresponding to a table, source, or
operator) in a single process. This allows us to funnel everything a
process is persisting through a single place for batching and rate
limiting.

Given a unique identifier, `PersistManager` will hand back a `Persisted`
token. This can be handed in (consumed) to create a timely operator that
persists and passes through its input.