let users specify the isolation level in Kafka sources
Diff passes the isolation_level property from a Kafka source down to
both the source and timestamper's librdkafka clients. Most users are
fine with the default, but there are some uncommon use cases where it's
useful to set this to read_uncommitted. Kafka reader config options that
aren't strictly required to be set a certain way for mz's correct
operation are better left configurable by the user.

Full UTs would require extending testdrive to write and abort
transactional messages, which was not worth doing for this change alone.
I did test manually, verifying that the config is correctly propagated
to two clients.

Closes #4915