coord: teach timedomains to also filter by timeline
With the addition of timelines, we can improve the time domain feature
of transactions. Previously (even before timelines were added), if you
had a CDC/Debezium consistency topic source in the same schema as an
epoch source, selecting from a source that had timestamps less than epoch
milliseconds inside a transaction would block forever because we helpfully
bumped up the chosen timestamp to be valid for all sources in the schema.

Prevent this by filtering timedomain ids to be in the same timeline as the
original query so users can have multiple timelines in the same schema.