Hopefully correct non-null requirement on aggregation parameters
We can convert a non-null requirement on an aggregate function into
a non-null requirement on its parameter only if we are sure that
won't affect the result of any other aggregation.
If there is non-null requirement on MAX(c1), for example, it is
implied that MIN(c1) won't be null either, so we can safely require
c1 to be non-null as long as there isn't any other aggregation
depending on any other column or a COUNT(*).