sql,coord: make PlanContext optional where not used
PlanContext previously had exactly one use: supply now() with a timestamp,
and only in oneshot SQL statements (i.e., not views). We had a lot of
machinery to copy PlanContexts around in places they were unusable anyway.

Make the pcx field of StatementContext private and optional, and instead
attach it to where it was actually being used: the OneShot lifetime. This
makes our data structures match more closely how the data is being used
(plan_current_timestamp no longer has to assume that the caller has
correctly set the wall_time).

Furthermore, we remove a bunch of calls to PlanContext::default that
were iffy anyway since they called into the system clock, and we should
be more pedantic about when things ask for the system time.