coord,pgwire: remove now() call during session PlanContext creation
Previously the only remaining call to SystemTime::now was in session's
creation of a PlanContext. In order to remove that call we need to
pass in the wall_time during transaction start. This requires starting
transactions to go through to coordinator, since it's now the only thing
with access to the configured now function (the single place where things
can get the wall time).

This exposed some problems in our transaction handling because the
session.pcx fn will now panic if there's no transaction. Parse and Bind
weren't starting their own transactions causing panics (Postgres starts
transactions in those messages, so we match them more closely). Similarly
failed transactions still need access to a PlanContext, so that variant
needed to be expanded.

Now the only calls to PlanContext::new are in start_transaction and
start_transaction_implicit. PlanContext::zero is only called by the
deserialization parse_item fn (and in tests). SystemTime::now is only
called in now.rs (except for one place in persist where it's not used
for time-related reasons).