persist: name streams externally as a string, make u64 id internal
It was going to be (doable but) annoying to map GroupId, which is an
enum of 3 u64s into our single u64 stream id. Additionally, sources
might end up with more than one stream (e.g. one for the regular row
output, one for the raw bytes coming in if we want to be defensive
against bugs in decoding, and one for timestamp assignments),
complicating this further. Now we can simply use something like:

    user-table-[schema]-[database]-[table]
    sys-table-[schema]-[database]-[table]
    user-stream-[schema]-[database]-[source]-output
    user-stream-[schema]-[database]-[source]-input
    user-stream-[schema]-[database]-[source]-ts

There is a bonus benefit that (future) debugging tools will be able to
print the name instead of a u64, which will be much easier if someone
wants to inspect what's been stored.

This keeps the internal u64 id in the batches we store to keep them
lean. The mapping happens once at stream registration, so there isn't a
performance concern for the translation.

Also, now that Id is an internal implementation detail, allow it to
derive Abomonation and remove all the various casts.