expr: fix nullability deduction of record fields
The type of a record field access expression like

    (record).field

depends on both the nullability of the field in the record type and the
nullability of the record itself. The expression is nullable if *either*
`record` or the field in the record is nullable. Previously we were only
considering the nullability of the field itself.

This requires adding an "optimization" to RelationExpr::typ to notice
when a RelationExpr::Filter is filtering out nulls on a column and mark
that column as non-nullable.

Fix #7170.