repr: fix memory unsafety in RowArena again
When we switched Row to use a SmallVec as its backing store (#3248), we
accidentally made RowArena unsound. It is NOT valid for the RowArena to
hand out references to a SmallVec, as the SmallVec might be using inline
storage, and the address of that inline storage will change if the
vector containing the SmallVec reallocates.

Fix the unsafety by forcing the SmallVec to spill to the heap, and
storing *that* vector in the RowArena.

In practice, the memory safety was only observed when packing wide rows
that contained small lists, but still scary to think how long this has
been present.

Fix #7168.
Fix #7169.