sinks,kafka: continuously update write frontier instead of only after commits
This ensures that we can compact source timestamp bindings even when no
data is flowing to sinks.

With the updated logic, while we're processing records of timestamp t
the write frontier is already at t. Which is okay because the Source
"since", which this holds back, is used to compact timestamp bindings <
t, not <= t.