materialized: reimplement telemetry reporting in terms of SQL
Rework telemetry reporting to use the SQL system tables instead of
Prometheus metrics. This avoids double-bookkeeping information for
information that can be derived from the system tables, like the number
of sources and sinks in the system and their types. Forcing all
telemetry information to come from the system catalog is also a nice
pressure to ensure we surface information to our users; if we want to
collect some metadata for telemetry, there's a good chance our users
want to be able to collect that metadata too.

I rewrote the version check loop to report new versions not just at
startup but also if a new version is discovered during telemetry
reporting at any point in the server's lifetime. As a result it's
probably easiest just to review the changes in telemetry.rs as new code.

Perhaps most importantly, collecting telemetry this way makes it much
harder to wind up with bugs like those below.

Fix #7137.
Fix #7181.