doc: propose beta features
The idea is that features that are ready for user testing but might
still have poor performance or stability get marked as "beta" in the
docs. This is purely about setting user expectations appropriately;
there is no opt-in required to use these features.

To start, PostgreSQL and Kinesis sources are marked as beta. This
overrides the old designation of Kinesis as "alpha".