dataflow,sinks: simplify TAIL sink
Before, there was a consolidate() and an arrange_by_key() with multiple
maps() strewn in-between. Now we have a single arrange_by_key(), which
also has the required consolidating effect.

Expanding a bit, the unrolled pipeline for TAIL, combining the
"envelope" and the sink code, was roughly as shown below. I'm taking
some artistic licence here, using Streams where the actual code as
Collections, to show the types more clearly, and other shortcuts:

let input: Stream<(Row, time, diff)> = ...
let enriched: Stream<(Row, time, 1)> = input
  .consolidate()
  .map(|(row, time, diff)| {
    let wrapping_row = Row::new();
    if (emit_progress) {
      wrapping_row.push(false);
    }
    wrapping_row.push(time);
    wrapping_row.push(diff);
    wrapping_row.push(row);
    (<some key>, (wrapping_row, time, 1))
  }
  .arrange_by_key();

sink::tail(enriched);

So the actual rows, along with their time/diff where put into wrapping
rows. And tail() would write out those rows.

Now, we lose the first consolidate() call, have one arrange_by_key(),
and then write out the user/vanilla rows straight from that, and create
the wrapping Row right in the sink:tail() code, without sending it along
Streams.

The assert! that we removed in sink::tail() was asserting on the diff of
the wrapping row. Now we straight get the vanilla rows, which can (and
should) have negative multiplicities.

We also move all of the code closer to the sink code itself, because the
keying by sink id is important for the implementation of the sink and
the knowledge of that should be isolated to the sink code.