test: use postgres v11
Configuration via the `-c` options in `mzcompose.yml` assumed that the
initialization in `setup-postgres.sh` was executed which is not true
anymore with newer upstream images (which results in us enabling ssl
via `-c ssl=on` but having no cert in place).

This moves everything to the image which does not separate concerns
between the test and image anymore, but it should be fine since we don't
have other uses of the image.

An alternative would be to configure everything in `mzcompose.yml` but
Postgres requires certs to be owned be the postgres user.