Extend expression cannonicalization
Remove redundant isnull predicates

Address item No. 1

Detect impossible condition when there is an IS NULL predicate and a
null rejecting predicate on the same subexpression

Extend NonNullRequirements transform

Extract non-null requirements from join equivalences.

Detect impossible conditions in NonNullRequirements transform when there
is an IS NULL predicate on a column that is required to be non-null.

Fix regression spotted in review

acceptable test changes