persist: only replay data once (instead of once per worker)
Before this commit, each of N workers would replay the entire persisted
data, creating N copies. This hasn't been caught before because all our
other tests are single-worker, so it happened to work. For now, do the
simple thing and replay all the data from worker 0 (and nothing from the
others). We'll make this smarter and balance the work out later.