testdrive: allow --temp-dir to be specified on the command line
If --temp-dir is not specified, a directory is created as before
using the tempfile crate. A PathBuf is passed downstream in both
cases.

We stash the tempfile::TempDir object, if used, so that it does
not go out of scope immediately and delete the temporary directory
that was just created.