repr: support chrono leap seconds
60 is a supported number of seconds in postgres, and in our datetime
parsing. However chrono expects seconds to be limited to 59 and allows a
leap second to be put into the nanosecond part. Have our validation logic
do that conversion. This should maybe happen in another place, but we have
other known problems and just preventing a panic is good enough for now.

We still have differences from postgres, but one less panic.

Fixes #7432
Fixes #6272