sinks,kafka: send consistency information with non-null key
Before, we were sending consistency updates with no key, which ends up
sending the updates with a NULL key. This doesn't work when compaction
is enabled for a topic.

Now, we set the name of the data output topic as the key. This should
not change the behaviour of existing use cases but will allow enabling
the aforementioned compaction.