dataflow-api: proof of concept for single-binary service boundaries
Key thing to note is that no crates other than materialized (and
coordtest) depend on dataflow. This is something that we'd want to
enforce in CI somehow so it doesn't accidentally rot.

This refactor was pretty straightforward because the abstraction
boundary of dataflow is already strong, with just a few exceptions
creeping in. The dep structure added by this PR makes it "impossible"
for exceptions to creep back in. We could even go so far as to disallow
deps from dataflow-api to mpsc and friends (once we clean up the ones
there now) to prevent that class of abstraction breakage from creeping
in.

Obviously tons to clean up here, this would never be merged like this.
Just showing how it might work to facilitate discussion.