persist: refactor Future snapshot to take an Antichain for its upper bound
Previously, the `snapshot` function for Future used an `Option<Antichain<u64>>` to
allow the user to indicate that they wanted to read all the data at times greater
than some lower bound, with no upper bound. However, we can express that intent
more idiomatically with an empty frontier as the upper, which corresponds more
naturally to the idea "the set of elements bigger than the desired times is empty"
which is the same thing as "there is no upper bound".