doc/syntax: Correct format_spec diagram and use it in all Kafka sources
Kafka is the only one of our source types that definitively supports every kind
of format, so we should document it on every page where folks go to get
information about Kafka sources.

I chose to show the primary formats as alternate options to the full
format_spec so that the main railroad diagram flow still calls out what the
current page is documenting.

In the future we should add some similar "complete" format specs that describe
the subsets of formats supported by other sources.