doc: Make Hugo more reliably pick up generate.sh changes
I've had a lot of trouble with Hugo apparently getting confused by the changes
from generate.sh, even with --disableFastRender --ignoreCache,

Moving the `rm` to just before the mv causes Hugo to not throw errors about
being unable to generate pages.

The second sleep and hot-swap seems to cause hugo to reliably notice that the
diagrams have been updated and that the site should be rebuilt.