sources,sinks: de-experimentalize RT timestamp binding persistence (AKA exactly-once sinks)
The title describes what it does underneath, but what it really does is
de-experimentalize exactly-once (aka reuse_topic) sinks with real-time
timestamping sources.

This also disables binding persistence for BYO sources, because those
are already "persisted" externally, in the BYO topic. We therefore
always pull the durability frontier to +Inf.