persist: support atomically writing to multiple streams
User table persistence will need this to ensure a SQL transaction is
persisted atomically. Modeling this as an atomic write is much easier
than adding a transaction protocol on top of persist and is also
sufficient from the coord side, so do that.

We already write Buffer entries atomically, so this is just generalizing
that and then a bunch of plumbing to expose it at the public API level.
The interface is a new `AtomicWriteHandle` which is constructed from a
set of existing handles from the same runtime. `AtomicWriteHandle` can
then be used to do atomic writes over any subset of the streams it was
initialized with.

Nemesis support is left as a followup.