testdrive: Create dedicated error report at end (#7560)
Without this an error in any test file causes Buildkite to display the final
test file's (currently S3) output, with two small lines at the end being the
actual error report. Having an expanded successfully-completed test file is
misleading.

This change also makes it so that that there is less to scan to see exactly
what the error report is, which is hopefully good for folks new to the system.