coord,sources: don't publish duplicate consistency information in BYO kafka connector
Session timeouts, together with consumer-group re-arrangement, combined
with the fact that we don't commit read offsets for the BYO consumer can
lead to cases where an existing consumer starts reading the consistency
topic from the beginning, after a connection outage.

We now protect against this by manually keeping track of the offset of
consistency records that we have processed so far, to filter out
duplicates.

Fixes #6586