coord: make metric scraper run on the main coord thread
The metric scraper is fast enough that it is arguably simpler just to
run it on the main coordinator thread.

A fair amount of code movement here is the result of moving `metadata`
into a field on `Scraper`, so that it persists between calls to
`scrape_once`, which causes some borrow checker woes.