persist: fix SELECT on persisted user tables
Broken in #7686 which separated the physical work of seal into step, but
forgot to call step after seal. Probably time to do the Tick TODO soon,
which is the real fix here.

Also certainly time for us to have some sort of end-to-end sanity test
that would have caught this. In the meantime, manually verified this
fix.