persist: de-flake allow_compaction
The underlying issue was that we were not reverting logical changes made to
in-memory state when those changes could not be durably committed to disk.

This issue was present ever since we added the allow_compaction command to
the nemesis tests and its unclear to me why we only noticed it recently.

The larger fix for these sorts of issues will be batching up multiple commands
and committing them all together, which should then remove the problem of
having to reason about committing / reverting each individual command.