coord,catalog: add status field/column for sources/views/indexes/sinks
The status is a transient field in the catalog, meaning it is not
persisted and will be re-created when materialized restarts.

The status can be Created or Available but all items have status Created
for now.

Future commits will change the status of items based on feedback from
the dataflow layer.