deps: bump to secure version of tar
Bump to a version of tar that isn't vulnerable to RUSTSEC-2021-0080.

For posterity, this vulnerability was not a security issue for us
because we only use the `tar` crate in a build script that downloads
packages from NPM, and NPM packages are allowed to execute arbitrary
code at installation time anyway.