mzcompose: Print some service logs when wait-for-<service> fails
If a wait-for step fails, the most valuable thing to do is to check the logs of
the thing that we didn't find the port for, but experience shows that that
isn't necessarily obvious, and it can be tedious especially in CI.

This implementation shows the most recent 20 lines of logs from a service that
mzcompose couldn't connect to. Even if it's not enough to debug the issue, it
at least is enough to point out that users should investigate those logs.

Internally, this improves the logic to use docker-compose applied labels
instead of the default naming scheme, which is robust to renaming/providing
explicit names for containers.