persist: call Indexed::step less frequently.
This commit teaches the persistence runtime to try to grab as many commands as possible
execute them in a sequence, and then call Indexed::step rather than calling it after
every single write and read. Doing so should amortize the cost of creating Future and
Trace batches over more commands.