coord,catalog,dataflow: update catalog item status based on feedback from workers
The dataflow layer only knows about dataflows for indexes and sinks
while coord/catalog also keep track of sources, views, and some more.
When updating the status of an index, we also update the status of the
source or view that it indexes, if it is the default (first) index.
Similarly, we need to update the status when dropping an index.

Additionally, we recursively update the status of (non-materialized)
views when any of their transitive inputs change.