materialized: remove most metrics from /status
Per @quodlibetor, it turns out folks weren't that interested in having a
set of human readable metrics on the /status page, and nearly everyone
who cares just wires up a Grafana dashbaord using /metrics. So, ditch
most of the metrics on the /status page, keeping only the version, query
count, and uptime. This removes a bunch of hairy Prometheus-wrangling
code.

I also refactored the query count into a separate metric, so we don't
have to go trawling through histogram summaries to derive it.