persist: fix duplicated record bug on restart
And enable listen coverage in nemesis, which is what found it.

Updating listeners when the data makes it into the trace was originally
my idea, but it was a mistake. The logic was that once a listener gets a
timestamp, it can reason that all data less than it has made it to the
trace's (not currently implemented) indexed access.

There were two problems with this. The first is a correctness bug where
everything in future gets replayed on restart as part of the snapshot
and then again via the listener once it makes its way into trace. The
second is that this needlessly adds latency to mz INSERTing into a table
and then immediately SELECTing from it.