Add (rudimentary) `bin/cloudbench` for running remote load tests, benchmarks, etc. in the scratch account.
`bin/cloudbench start` creates a number of remote clusters on a set of Git revisions.
On each of these clusters, a user-specified Python script is run, which must print
a CSV-formatted list of metrics with headers (and nothing else) to stdout. Cloudbench
returns control to the user after launching all the clusters.

For example,
```
bin/cloudbench start -n 3 --profile basic --revs 'HEAD,HEAD^' misc/python/materialize/benches/bare.py -t 60 -n 60
```
launches six machines, each of which runs the given script (`bare.py -t 60 -n 60`).

The `basic` profile only launches one machine per cluster; a follow-up PR will introduce a `confluent` profile
that starts two machines, one of which runs the provided benchmark script, and the other which runs kafka, schema registry, and so on.

Note that each script may print multiple CSV records, which we refer to as "trials".

After starting all the required machines, the start script prints out a benchmark ID which can be used to address that set of machines in calls to `bin/cloudbench check`, as described below.

When the user wants to check that the scripts have completed and get their output if so, he or she may run
```
bin/cloudbench check <bench_id>
```.

This command checks every one minute for completion of the benchmark scripts on all machines. Once they have all completed successfully,
it downloads the CSV outputs, adds a few metadata columns (trial number, instance index, and git rev), and prints them to stdout as one CSV file.

As stated in the PR title, this is still rudimentary, and there is a lot of possible future work:
* Better error-handling (currently, `bin/cloudbench check` simply aborts if any of the remote scripts failed)
* Implement the `confluent` profile as described above
* Have the machines store their scripts durably, so the results are not lost when our cleanup CI script reaps them,
* Make it possible to test revisions from before this commit was introduced, by checkout out the relevant scripts from `HEAD` on the remote machine, or uploading them from the local machine
    (this is done now with the actual benchmark script, but we should do it for all files that are necessary to use this infrastructure).
* Create some convenience scripts to analyze the results CSV and