coord: implement sql linearizability
Use a new Timeline struct to keep track of linearizability per
timeline. When we need to create a read or write timestamp, use the
Timeline struct to ensure we are producing a timestamp that meets that
guarantee.

Change tables to use this as their linearizability tracker too. Tables
are no longer special in `determine_timestamp`. Previously, a table would
fetch a read timestamp (`now()`) and also trigger closing inputs up to
that time, which would thus always cause the table Peek to wait until
the closing message had been processed by all dataflow workers. Now we
can choose table read timestamps using the same heuristic as all other
things, and still provide linearizability.

The logic in `advance_local_inputs` had to divide up responsibilities
with the new Timeline struct. Things that have to do with correct
timestamp selection go into the struct, and things having to do
with timestamp closing policy remain in Coordinator (`closed_up_to`,
`need_advance`). This does mean that table INSERTs are the only caller of
`get_write_ts` in the new struct, but we at least have a safe way to do
that if we are ever need to support INSERTs for other timelines somehow.

timestamps-debezium-kafka.td had an existing linearizability violation
where it emitted a row and then in a later query omitted that row from
a join'd view since one of the sources hadn't advanced yet.