S3: support FORMAT CSV WITH HEADERS (col, ...)
This supports headers in CSV objects in S3. Currently this requires that users
specify the header column names, and errors in header column names don't appear
until the first SELECT query executed against a materialized source.

Work is ongoing to make the column names optional and to make the validation
happen as part of purification, but this is a nice simple change that unblocks
the most important parts.