S3: Eagerly exit SQS listener and Object downloader on shutdown
Before this change, it was guaranteed that SQS readers would successfully
process, and delete from the SQS queue, at least 1 message before they could
notice that the dataflow should shutdown.

This happened because of the way propagation of dataflow shutdown was
communicated:

* The S3SourceReader (which is the actual object that exists inside a dataflow)
  holds an mpsc Receiver, responsible for picking received *data* off of the
  `download_objects` queue.
* `download_objects` holds the sender for the dataflow. It also holds a receiver
  that is used by all of the SQSListener and BucketScanner tasks.

The dataflow notified the `download_objects` task that it shut down by dropping
its receiver, causing `tx.send`s to fail. Once the download_objects task failed
to send a message on the dataflow mpsc sender the task notifies upstream tasks
-- specifically for the case of the bug that we care about the `sqs_listener`
task -- by shutting down *its* mpsc receiver.

This all is sort of fine, except for the fact that download_objects could only
attempt to send data to the dataflow (and thus learn that it should shut down)
once it had received a message telling it that there was an object that it
should download. In the case of SQS queues, reading the message with the
"download this object" notification is destructive. So it's impossible for SQS
queues to shut down in this scheme without destroying at least one message
about what should be downloaded.

This is likely to be primarily a papercut, it should only be observable for
users who are doing exploration and constantly creating and recreating views
listening to SQS queues. Of course, these are exactly the people most likely to
bounce if Materialize seems to lose data in interactive use.