pgcompat: add pg_catalog.pg_get_expr
pg_get_expr is meant to convert the textual version of
pg_node_tree data into parseable expressions. However, we don't
use the pg_get_expr structure anywhere and the equivalent columns
in Materialize (e.g. index expressions) are already stored as
parseable expressions. So, we offer this function in the catalog
for ORM support, but make no effort to provide its semantics,
e.g. this also means we drop the Oid argument on the floor.

Fixes #7785