persist: unbreak nemesis with trace compaction.
We previously merged trace compaction without adjusting nemesis / trace snapshot
to be able to account for differences in data due to advancing the since frontier
and logical compaction. This commit teaches Indexed to advertise a since frontier
as part of its snapshot, and teaches nemesis to respect that since frontier when
comparing the expected vs actual values of data when reading a snapshot.

It was not clear how to do the same for read_output commands so this commit also
temporarily disables those in the interest of getting nemesis unbroken as soon as
possible.