row: Manually implement Clone
By default, Clone of SmallVec uses a non-optimized copy method for the
backing allocation. The reason for this is that Rust currently doesn't
support specialization.

Instead of relying on the derived Clone implementation, provide a manual
implementation that calls into SmallVec::from_slice, which copies the
backing data in one operation.

Fixes #7927.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>