persist: version the serialization of BlobMeta
Because we're temporarily using abomonation as our serialization format,
any changes to the structure of the BlobMeta struct (or the many structs
included in it) will change our on disk representation in a non
backwards-compatible way. The "META" key in Blob storage is the very
first thing read by persist on startup and the encoding used for it is
also perfectly predictive of the encoding of BlobUnsealedBatch and
BlobTraceBatch (the other two things we serialize).

This will allow us to gracefully detect changes at startup and react to
them. MZ is not (yet) a system of record, so in the short-term
(persisting mz_metrics) we are free to simply delete the data. In the
medium-term (fast restarts for kafka sources), we'll make an effort to
decode old data. In the long-term (1.0), we'll have
backward-compatibility guarantees.

A followup PR will close this loop and, on startup, wipe the entirety of
persisted data if an unsupported version is detected (safe for the
reason described above), before initializing a new copy and proceeding
on.