coord: save each user table's persist name in the catalog
And use it on startup if it's present instead of generating one from
scratch. This fixes two bugs: either of renaming a user table or
changing the user table persist name generation logic
(PersisterWithConfig::stream_name) would disassociate and orphan the
previously persisted data for that table.

System tables don't get persisted in the catalog in the same way, so if
we change a persisted system table name, a persisted system table id, or
the system table persist naming, we'll need to clear existing persisted
data (which is safe because the system table persistence is best
effort).

Fix manually verified by renaming a persisted user table. Test is being
tracked in #7900.