sql: support ORDER BY in aggregates
For aggregate functions that can support ORDER BY, always map them to a
Record with optional entries that are the columns to be ordered on. Use
a new Func variant to disambiguate which aggregate functions take ORDER
BY columns and which don't. We always go through the ORDER BY work
even if it's not specified by the user because it simplifies a lot of
code, and these functions are all already anti-patterns, so we're not
especially concerned about optimizing them. In general, users should
not be sending large numbers of datums to them, so the extra work is
hopefully not noticeable.

Fixes #2415