sql: add information_schema._pg_expandarray
Add the information_schema schema, as this is its first function.

Add a new Func variant that's similar to table funcs, but directly
exposes a relation expr and scope. Add a function that can produce the
appropriate expr given a SQL statement, allowing us to implement this
function in a way similar to Postgres.

Although we have added support for _pg_expandarray, it's still not yet
enough to support PgJDBC, which needs #1546. However we can add the
FROM-side support for this function now, and separately implement 1546
which will turn this on.