release: Add a script to create the release qual dashboard links
Creating the links is one of the most annoying parts of the release process --
you have to go through a UI and click various things, and then be sure that you
haven't done the same thing twice, and feel confident that Grafana correctly
updated the URL as you were clicking around.

It is also one of the easiest ways to do something that makes future
comparisons difficult: the best thing to do when creating the links is to
create an explicit time range. The default on all the dashboards is to use
something relative to "now", and "now" changes as time in the universe marches
onward. That is an additional annoying step, which is rarely done.

This turns the whole thing into a single command, which is easy to verify.