upgrade: Add an "upgrade-from-latest-to-current-source" test
This is similiar to, but distinct from, the revert that happened in #7839. We
do not expose `latest` as a create-in or check-from target -- i.e.
`create-in-latest_version*` doesn't do anything. Instead, we *do* always test
that the current source can build from the latest released version. This
specific test would have caught #8202, and generally knowing that the next
version of materialize will load the catalog of the previous version is a good
base case to test.