sql: namespace item imports of sqlfunc! macro
Previously if the macro was invoked twice in a file the second
invocation would error out with duplicate import errors, since there is
no hygiene for items.

We now create a module with the same name as the provided function that
contains all the definitions and then pub use the consturcted struct out
of it.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>