dataflow,persist: wire persistence configs, this enables previously added persistence code
This wires up the user-facing option that we added in a previous commit
to enable the previously inactive persistence code paths.

The "persistent_name", as stored in the catalog and visible in the
introspection views is used for the collection that stores the actual
source data. We append "-timestamp-bindings" to form the name of the
persistent collection that stores timestamp bindings. The rationale
behind this is that possible consumers of the persisted source data need
only concern themselves with the actual source data while the timestamp
bindings are more of an internal implementation detail.

Additionally, we seal both the timestamp bindings and upsert state
collections when persistence is enabled.

Note: we currently don't add seal errors to the differential error
collections, meaning a seal error will not "wedge" an index. Instead, we
only log an error. Seal errors are transient: we don't lose data when a
seal doesn't succeed, all that happens is that downstream progress will
be blocked until we can seal again.

Running the `avro-upsert` benchmark on a Ryzen 3700X with 64GB of
memory, we get this with persistence enabled:

./bin/mzbench --size medium --benchmark-id hello --num-measurements 1 avro-upsert
git_revision,num_workers,iteration,seconds_taken,rows_per_second
None,7,0,397,50378
None,5,0,553,36166
None,4,0,598,33445

While we get this without persistence:

git_revision,num_workers,iteration,seconds_taken,rows_per_second
None,7,0,26,769231
None,5,0,34,588235
None,4,0,42,476190

The large difference comes mainly from how the seal/persist operators
block on the seal/write futures, which means the timely worker threads
are blocked for long stretches of time. Follow-up PRs will make the
seal/persist operators non-blocking.