dataflow,persist: wire persistence configs, this enables previously added persistence code
This wires up the user-facing option that we added in a previous commit
to enable the previously inactive persistence code paths.

The "persistent_name", as stored in the catalog and visible in the
introspection views is used for the collection that stores the actual
source data. We append "-timestamp-bindings" to form the name of the
persistent collection that stores timestamp bindings. The rationale
behind this is that possible consumers of the persisted source data need
only concern themselves with the actual source data while the timestamp
bindings are more of an internal implementation detail.

Additionally, we seal both the timestamp bindings and upsert state
collections when persistence is enabled.