expr: return type of sql funcs can be more than one token
Previously `sqlfunc!` was using a `$ret_ty:tt` metavariable to capture
the possible return types of the provided function. This works for
simple types but fails for more complex ones like `DateTime<Utc>`.

This patch uses a proper type typed metavariable `$ret_ty:ty` to capture
the types and splits the macro processing in two stages.

Splitting the stages explicitly is required to avoid having the `ty`
typed metavariable capture more tokens that we want.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>