coord: use all indexes per id in timestamp determination
Previously determine_timestamp would only use the first index per
id. If there were multiple indexes with different sinces, and the
second (not included) index had a higher since, it would not bump the
timestamp forward to that later since. Then later on if the non-fast path
branch of sequence_peek was taken, the dataflow builder would import
both indexes. When ship_dataflow saw that, it would panic because the
requested timestamp was before one of the imported indexes.

This was easily triggerable if a transaction was started which would
prevent compaction of the first index, but not of the second. Further
statements in the transaction could then use an index that had not had
its compaction held back.

Fix this by including all indexes from an id.

Fixes #8241