persist: fix bug making compaction too aggressive when under high request load
HACK: This rate limits how much we call step in response to workloads consisting
entirely of write, seal, and allow_compaction (which is exactly what we expect
in the common/steady state). At the moment, step is too aggressive about
compaction if called in a tight loop, causing excess disk usage. This is exactly
what happens if coord is getting steady read traffic over indexes derived
directly or indirectly from tables (it continually seals).

The downside to this rate limiting is that it introduces artificial latency into
filling the response futures we returned to the client as well as to updating
listeners. The former blocks literally no critical paths in the initial
persisted system tables test. The latter holds up reading from the system tables
being persisted, which is considered an acceptable tradeoff for the short-term.
In the long-term, we'll do something better (probably by using Buffer).