persist: add await_frontier() operator
Stashes data until it is no longer beyond the input frontier.

This is similar, in spirit, to what `consolidate()` does for
differential collections and what `delay()` does for timely streams.
However, we are not working with differential collections and `delay()`
deals with changing the timestamp while the behaviour we want is to wait
for the frontier to pass. The latter is an implementation detail of
`delay()` that is not advertised in its documentation.