coord: improve timedomain id txn verification to include indexes
The set of indexes could have changed after a transaction started, so
we need to verify that there's no new indexes that the dataflow builder
might use. This is hopefully a temporary workaround until #8318. It is
a bit unfortunate because it's a side effect from another sql session,
so we have gotten further away from serializability.

See #8318
Fixes #8241