doc/user: avoid the term "Materialize node"
Since Materialize is a single-node system, I think it's a bit confusing
to use the term "node" to mean "Materialize server". It's familiar to
those of us used to CockroachDB terminology, or other distributed
systems, but I don't think it's immediately obvious to most users.  And
when we *do* support horizontal scalability, per the platform vision, I
think we're likely to say "Materialize worker node" or "Materialize
compute node" rather than just "Materialize node".

So, this commit rewrites all use of the term "node" in the docs to use
either "server", "deployment", or "catalog" instead, depending on the
context.