pg-cdc: handle unchanged toast values during updates
Due to misunderstanding of the replication protocol the postgres source
was previously was over-restrictive with respect to handling unchanged
TOAST values.

In the general case, we can not ingest the placeholder value that the
replication protocol uses to indicate that a TOASTed column is
unchanged.

However, when the table in question has `REPLICA IDENTITY FULL`, which
is what we recommend anyway, the replication protocol will only use the
placeholder value for the *after* value of a row during an update but
for the *before* value it will include all the data.

Therefore, when we encounter an unchanged TOAST value placeholder in the
after row, we can grab the previous value from the before row and send
that in the dataflow.

Fixes #8362

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>