coord: verify prepared statements on use
Previously we would cache the StatementDesc and assume it didn't
change. Add some new APIs (and a forced trip through coord) to do the
re-verification. Use the catalog version as an optimization, allowing
us to ignore checking because we know it's still valid.

While testing this some panics were found in the pgwire Describe path
which assumed a transaction had started when it hadn't.

Fixes #8397