persist: don't block replay operator construction on snapshot iter
We were previously waiting for a snapshot to be entirely iterated (and
materialized in memory) as part of dataflow construction, which is
expected to be fast. Instead, iterate the snapshot when the operator
gets scheduled.

Still TODO is making snapshot iteration data parallel so we can use the
other workers for this.