persist: maintain pending_writes in a hashtable
Previously, we were maintaining pending_writes in a vector that was then transformed
into a hashtable in drain_pending. This approach adds a little bit of latency to
each write call (which previously amounted to a memcpy more or less) but in return
we get a substantial latency reduction in `drain_pending`. This commit also lays the
groundwork for `drain_pending` to be copy-free.

This change doesn't register as a meaningful performance improvement in the
indexed_write_drain benchmark. Before this commit:

indexed_write_drain     time:   [495.17 ms 497.63 ms 500.15 ms]
                        change: [-0.7764% -0.0550% +0.6712%] (p = 0.88 > 0.05)
                        No change in performance detected.
indexed_write_drain_sorted
                        time:   [328.01 ms 330.52 ms 333.04 ms]

With this commit:

indexed_write_drain     time:   [494.97 ms 497.34 ms 499.79 ms]
                        change: [-0.7403% -0.0583% +0.6318%] (p = 0.85 > 0.05)
                        No change in performance detected.
indexed_write_drain_sorted
                        time:   [329.70 ms 333.20 ms 336.88 ms]
                        change: [-0.5895% +0.8097% +2.0959%] (p = 0.23 > 0.05)
                        No change in performance detected.