repr: support BC and AD era identifiers for most timestamps
Add support in the date/time parser and formatter for BC/AD identifiers,
trying to support as many cases as supported by Postgres. Most of the
challenge and complication here comes from how flexible Postgres is
with these identifiers: they can be placed almost anywhere in a datetime
string, as long as they come after the year.