dataflow: allow source operator to emit a secondary output stream
The main intended use case for this is to allow sources to additionally
(and optionally) emit timestamp bindings, so that we can persist them.

We need to use the low-level operator builder instead of the
higher-level timely `source()` operator. The code is almost equivalent
to what `source()` uses internally, with the addition of the secondary
output.

This will be used by the persistent Kafka source work to persist
timestamp bindings.

Note: neither SimpleSource-based nor SourceReader-based sources use this
right now. We plan to implement it for SourceReader-based sources in a
next commit.