persist: WIP for how we might move listeners off the hot path
The model of returning reqs from step is the same as the way compaction
reqs work in #8472. Something that's odd is that compaction reqs can be
executed in any order (in fact are intended to execute concurrently),
but for correctness to hold, listeners must be executed in the same
order as they are generated. It seems like a bit of a wart to lump them
both in under MaintenanceReq.