persist: move decoding to the consumers of listen streams.
Previously, StreamReadHandle::listen would first decode the records being sent to
listeners before invoking their provided callback on each decoded listen event.
This was a fairly nice and ergonomic API, however it had the unfortunate side
effect that all of this decoding work was performed by the persistence runtime
thread.

This approach changes the API of StreamReadHandle::listen to one where users
provide a listen function in terms of the encoded data. In the persisted_source
operator, which is the only place this is currently used, decoding now happens
on the timely worker thread, not in the listen function.

Note that a user could still perform CPU-intensive work like decoding in their
provided listen callback. It's unclear at this time whether we need to do something
else to guard against that.

None of the existing benchmarks were affected by this change, however, tests on
a separate branch writing 1 million 100 byte records to a table showed a ~220 ms
reduction in write latency.