interchange: fix protobuf schema bug (#8528)
Previously, we would try to grab the first protobuf message from the
first file descriptor in a set to try to use as the top-level message for
decoding. However, not every file descriptor contains a message! This
changes the code to look through all available file descriptors instead,
returning the first one found.