dataflow,coord: move persisted_name from SourceConnector::Local to SourceDesc
Going forward, real sources will also need to know a persistence name.

Plus, it seems more natural to put it in the SourceDesc, since this is
what the coordinator creates on the fly when sending a dataflow
description to workers. The source connector is created in the parser
(at least for real sources), at which point we don't yet know a
persistence name.