optimization: more efficient FoldConstant (#8609)
Copying data into materialize from a stdin is slow (~10s for 100MB), with time mostly spent in the optimization stage, specifically with a lot of time spent in the FoldConstant optimization. The current way in which we consolidate the constants is memory-inefficient and causes a lot of time to be spent in memcpy. Replacing this with differential_dataflow::consolidate::consolidate brings a noticeable improvement in this benchmark.

Before:
time gives: real 0m15.333s
After:
time gives: real 0m7.158s