dataflow: allow injecting initial offset in add_partition()
This adds the plumbing that will allow injecting offsets that we
restored from persistence, which we will add in a later commit.

This also removes the blanket add_partition(PartitionId::Kafka(0)) calls
we have in the source initialization code and relies instead on adding
the partitions when getting the command to enable timestamping from the
coordinator.