sql: allow NOT ENFORCED keys for Kafka Upsert Sink
https://github.com/MaterializeInc/materialize/issues/8200 has shown that
there are use cases in the wild where our optimizer cannot derive
a unique key of a sinked relation even though as a user we can see that
a given key should be a unique key.

This introduces the NOT ENFORCED escape hatch, which allows disabling
the key check.

This does not mention the syntax in the documentation or in error
messages because it is a potentially dangerous thing to use it.
Incorrectly using it to enforce a key that is not unique means that
materialize can panic at runtime.