persist: in conditional_seal() make types of sealed collection independent
Before, the types requires that the types of the collections we are
sealing correspond with the types of the stream whose frontier
information we use for sealing. This is too restrictive, for example in
the upsert operator we want to seal both the upsert state and timestamp
bindings outside the upsert operator, at which point the Streams have
different types from the collections we use to maintain state.