persist: remove next_blob_id
The latest update of nemesis (#8649) caught a bug where we were
attempting to reuse an id (and thus a key). We could track the bug down,
but for several reasons it's better to just remove next_blob_id
entirely.

- They are different than the other information in META in a critical
  way: everything else in META is revert-able to a previous state, but
  we can't revert next_blob_id because they might have been used.
  Indexed's revert has to jump through some hoops to deal with this,
  adding unnecessary complexity.
- Most of META itself is representable as a fixed number of collections
  (in the differential dataflow sense). This might be exploited in the
  near future to incrementally maintain META in storage. next_blob_id is
  one of the few (all fixable) exceptions to this pattern.
- Persist keys are meant to be opaque, but by not looking opaque, they
  subvert that expectation.

The big con is that keys are no longer deterministic, which is mostly a
hassle in unit tests.