sql: prohibit creating map[], list[], char list
We previously only performed checks for verboten complex type
combinations when planning literal values, not in all contexts
that expressed those types, e.g. as columns or as user-defined
types.

This commit adds those checks. Note that this is a breaking change
w/r/t char list, which was previously a valid custom type, but will
no longer let Materialize boot if it's defined.