s3: generate partition ids for each object
Previously all S3 objects were given one by one to the dataflow and used
MessagePayload::EOF in order to mark the beginnings and ends of each
file.

With this change each S3 object is tagged with a unique partition id
that is computed using a hash of its bucket and path.

While MessagePayload::EOF is still used to demarcate the end of each
individual stream, each object is now sent to a randomly picked worker
in order to decode in parallel.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>