coord: force static views into the system timeline
Previously, a transaction started on a static view would have had no
timeline, but the timedomain code opted it into the system timeline,
thus pulling in most sources/tables. But since the original timeline
variable was None, none of the linearizability tracking (or now() capping
in this case) would trigger, causing the table SELECT to try to wait
for end-of-time.

To fix, map queries with no timeline into the system timeline.