persist: in persistent_source(), build consistent operator graph
Before, we were building different shapes of graphs depending on whether
taking the snapshot/listening would succeed or not.

Now, we handle the error cases internally, in the operators. To the
outside world, the shape of these operators is always the same now,
regardless of error cases. This deterministic graph construction on
multiple workers if important for timely dataflow to work correctly.

Fixes #8687