persist: use multi-threaded timely dataflows in nemesis
This is another step toward adding concurrency to nemesis. Previously,
all dataflow work was run in a single worker deterministically stepped
by Direct. Instead, run it now in two background threads. This makes it
more likely that things like dataflow construction and dataflow
execution will be run concurrently with writes and other persist
traffic. It also starts to model the way persist will be used in the
platform world a little more closely (where the number of dataflow
workers emitting a persisted collection doesn't necessarily match
however many threads are driving the input side of it).

This finds #8687/#8739.

The new structure makes it less obvious how to do the previous trick we
had in nemesis where a seal would block until the dataflow had caught up
to ensure that subsequent ReadOutput steps would have an interesting
amount of output to read. Most of the complexity is porting the spirit
of this to the new world. I initially tried to remove it, hoping that
running dataflow in the background might keep it sufficiently caught up
and the complexity might not be needed anymore, but it turned out to be
pretty necessary to getting nemesis to find #8606 (which it does readily
with the following approach).

What we do now is make the seal req block on the dataflow catching up,
but only if the actual seal part succeeded (so we don't wait forever and
hang if the seal failed because storage was down). Combined with the
MAX_OUTSTANDING in Runner, we end up with a nice behavior where seals
and read_outputs are allowed some amount of concurrency, but eventually
we'll block issuing any new reqs (including read_output) until the
dataflow has caught up to the seal.