persist: prevent MemBlobCore from being used by outdated MemBlob refs
This fixes a regression that was affecting the nemesis tests where async
fetches happening in background threads could race with a close and
re-open of MemBlob and then incorrectly still affect the newly open
MemBlob generation though the handler for the previous MemBlob
generation.

This is really only a problem for tests, but it's a common pattern in
tests to model restarts, so it's worth getting right.