persist: address TODO to handle the operator error output in nemesis
Basically, we make no guarantees once an error has been output, so
nemesis allows anything. Technically, this could have run the entire
production path logic of filtering only the errors before the timestamp
the data is begin read at, consolidated those, and only consider it an
error if the resulting set was non-empty, but persist doesn't do
anything with retractable errors yet, so that seemed unnecessary.

Split out from the WIP nemesis concurrency mega-PR because this is
something we could have been doing all along and I'm looking for ways to
break that up.

The bigger change here is making the output of the PersistedSource
operator be a Result stream instead of the previous two streams (one for
data and one for errors) mostly because it was less plumbing in nemesis
and we'd been talking about doing this anyway. Lmk what you think, it's
also straightforward to revert that bit and handle two streams in
nemesis if we think this is premature.