dataflow: make decoders oblivious of position metadata
The decoding stage is responsible for transforming a timely stream of
`SourceOutput` messages into a timely stream of `DecodeResult`.
Previously, depending on the `push_metadata` flag, the format decoders
would push the position metadata into the decoded row. However,
`DecodeResult`s have a `position` field that carries the position
metadata anyway.

In this patch format decoder implementations are made oblivious of the
`position`[1] and `push_metadata` information and never produce this
information in the output row.

Then, the downstream envelope stage of transforming the `DecodeResult`
stream into a pure `Row` stream can choose whether or not it makes sense
for the envelope to append the position metadata into the output `Row`.

Previously the rule for the `push_metadata` depended not just on the
envelope type, but also on the format. Specifically, if the format was
Avro the position metadata was being omitted. This chagne makes pushing
the metadata into the final row depend only on the envelope type, which
introduces a breaking change.

After this patch all envelopes except `DEBEZIUM` and `MATERIALIZE` will
forward the position metadata into the final row.

The users that will observe a breaking change are all users using an
Avro source (this does not include AvroOCF) with an envelope of `NONE`
or `UPSERT`.

[1] The avro format decoder still needs access to position information
because it's intermingled with debezium logic. The property of never
pushing the metadata into the output Row still holds though, which is
sufficient for this patch. Once the debezium logic is fully implemented
on top of Rows the Avro decoder will be oblivious of position
information.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>