persist,dataflow: forward persistence errors from persist/upsert operators to differential error collections
It is not ideal that persistence errors end up in the same differential
error collection as other errors because they are transient/indefinite
errors that we should be treating differently. We do not, however, at
the current time have to infrastructure for treating these errors
differently, so we're adding them to the same error collections.