dataflow: Pull Permutation planning up in delta joins
This moves the planning of the Permutation to the MIR->LIR translation
phase instead of planning it at run-time.

At this time, the stage plans for linear and delta joins are not equal
anymore. The main reason for this is that in linear joins we still have to
rely on run-time permutation information because we might be getting a
stream or arrangement as inputs.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>