persist: in persist() operator, don't block on write futures
Before, we would synchronously block on the write future using recv().
This meant that we block the timely worker thread far longer than we
need to.

Now, we convert it into a "real" Future and poll it asynchronously.

We are also removing an inefficiency, where we would wait for the
frontier to advance beyond the timestamp of a write future, before
blocking on it. Now, we just hold on to capabilities for the timestamp
of the written data until the write future is done, making the operator
both simpler and more aggressive in allowing the frontier to advance,
because the tracking is now more fine-grained.

With this change and the change to make conditional_seal() non-blocking,
we get these benchmark results for `avro-upsert` on a Ryzen 3700X with
64GB of ram, with persistence enabled:

./bin/mzbench --size medium --benchmark-id hello --num-measurements 1 avro-upsert
git_revision,num_workers,iteration,seconds_taken,rows_per_second
None,7,0,25,800000
None,5,0,34,588235
None,4,0,43,465116
None,2,0,84,238095

This is the same as the result without persistence. However,
`avro-upsert` is not correct because it only measures how many messages
we ingested from Kafka, without measuring the time to go through
persistence. That is, we don't measure when the data would be available
to query.

Closes #8813