persist: turn Blob trait into an async trait
This commit only changes the functions of the trait to being async and
does not touch any of the implementations to actually make them async.

As a result, the majority of this diff is adding .await points on all
the call sites that use blobs but stops the "asyncification" when
encountering the public interaface of another component. So for example
`BlobCache` is not change to be async itself in this patch, it just
wraps blob calls with `block_on`.

This allows us to change persist into being async step by step by
removing one layer of `block_on` at a time.

to being async and does not

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>