transform: Avoid a clone for pretty error messages
The fixpoint transformation can fail, and in this case it produces a
pretty-printed error message. The issue at hands is that the implementation
unconditionally creates a clone of the input relation, which is only used
to show an error message, if there is an error.

While it is nice to have good error messages, they should not come with a
high runtime cost. For this reason, remove the original expression from
the message and only show the transformed relation.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>