repr: Add RowAllocator
The current mechanism to avoid row reallocations is to extract a copy
of a row while leaving the allocation around for later. This is problematic
because it potentially leaves an allocated row around that will only be
used much later or never again.

The RowAllocator tries to mitigate this by not retaining an allocation but
remembering the maximal length of the previously produced rows. Once
borrowed, it allocates a row of the desired length. The row then can be
extracted, which also updates the length information.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>