dataflow,persist: persist timestamp bindings and use them to restore source offsets
This does two things: 1) emit and persist timestamp bindings from
SourceReader-based sources and 2) use the persisted bindings upon
restart to derive the latest partition offsets and seed the timestamp
bindings histories.

The new code is only active if `Some(PersistentTimestampBindingsConfig)`
is passed.  For now, all call-sites pass `None` and we ignore the
returned timestamp bindings that we can use to track the persistence
frontier, which we can in turn use with `conditional_seal()` to seal up
the backing collection.

We will wire this up to a user config for persistent Kafka upsert
sources in one of the next commits.