persist: fix conditional_seal() operator
With `conditional_seal()` only one operator is active and doing any
sealing. The other operators just pass through data.

Before this fix, we held on to a capability on all workers, meaning that
we would block downstream frontier progress for non-active workers. Now,
we correctly only hold on to a capability for the active worker.