coord/persist: teach persist to handle varying encoding versions.
This teaches persist to, on startup, check what version of encoding format we
previously used to write down persisted data.

If the encoding format is less than what the current version of Materialize can
support, we delete all of the persisted data. Otherwise, if the encoding format
is greater that what the current version of Materialize can support we error, and
refuse to start up.