repr: move &self Row methods under RowRef and use Deref
Make `RowRef` the canonical representation of a borrowed Row. All the
methods on Row that had a `&self` receiver have now been moved under
RowRef and are through a `Deref<Target=RowRef>` implementation.

It is necessary to mark RowRef's representation as transparent in order
to safely transmute from `&u8` to `&RowRef`. The safety argument is
exactly the same as stdlib's for the equivalent cases of `PathBuf` ->
`Path` and `OsString` -> `OsStr`.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>