persist: pull try_set_meta up to Indexed
Also introduce apply_unbatched_cmd, the first of two wrappers for
try_set_meta that will simplify its usage.

Requests are split into two types: _unbatched_ and _batched_. An
unbatched command is run entirely by itself (the applied state has just
been written to durable storage, then the command is run, then the
resulting state is immediately written to durable storage). A batched
command is applied to the machine state, but instead of immediately
serializing the state to storage, we buffer the command response in
Pending. Any other batched command can also be run and similarly
buffered in Pending. Then, the next time we get an unbatched command (or
`step` is called), all pending batched commands are made durable at once
(and responses filled, listeners updated, etc). This is a performance
optimization to amortize the cost of writing to durable storage across
many of those requests. The most common requests (write, seal,
allow_compaction) are all _batched_ to exploit this. All unbatched
commands are expected to be relatively infrequent (to avoid excessive
barriers in our pipelining).