persist: clean up operator error_stream tests
These tests currently rely on internal impl details and so the error
messages in the tests often need to be updated in response to unrelated
PRs. Attempt to combat this by shutting down the runtime instead of
using unreliable storage. The former error happens much earlier in the
call graph, so hopefully it will need to be updated less often.