testdrive: do not retry COMMIT and ROLLBACK on error
Since the transaction is aborted if the first COMMIT fails, subsequent
attempts to retry that statement will result in a success, which is
misleading, since a new transaction context is now in effect.

This causes testdrive to output a bogus and confusing 'query succeeded,
but expected error' message.

Fix by adding a "should retry" check to the code that executes
SQL statements that are expected to fail and have the check return
false for COMMIT and ROLLBACK.