persist: switch Description<SeqNo> to Range<SeqNo>
Our usage of Description for data timestamps in trace batches is
correct, but somewhat just because it was there, we also used it for
declaring SeqNo bounds of unsealed batches. In that role, it's correct,
but unnecessary. First, we never compact anything on a SeqNo basis, so
the since is always "zero" (the single element antichain of the minimum
SeqNo). Second, SeqNos are totally ordered and always will be, so
there's no need for the lower and upper bounds to be Antichain (as
Description hardcodes).

In theory, Description is just a superset of the information we need to
represent and our usage of it was correct. In practice, the Antichain
caused some bits of code to be unnecessarily complicated and we probably
weren't being defensive about erroring on non-"zero" sinces.