persist: disable background test when logical_compaction_window is off
When logical_compaction_window is off, the amount of data in the metrics
system tables being persisted as part of our background test would grow
without bound and use an unbounded amount of disk. Given that this test
is supposed to be transparent to the user, we disable the test when this
option is used.

Touches #8184