persist: don't leak blobs of destroyed streams
This also starts moving blob deletion out of the hot path and into
background maintenance by modeling pending deletions as state. A req/res
protocol for sending these to Maintainer is left as a followup.

This could also possibly be the basis of a protocol to ensure we don't
leak blobs. Currently this can happen if a new unsealed or trace blob is
written, and then storage becomes unavailable until the process restarts
so the key never gets written into META. We could instead first write
the key into pending_deletes and make that durable, then write the blob
at that key, then atomically update META with the reference to the key
and removal from pending_deletes. We don't do that now because it would
add a barrier on an extra META write in many places.