demo/billing: deflake
At some point in the past, we started requiring the message name in a
protobuf source to start with a leading dot. This broke the billing
demo, which was never updated for the new syntax. This commit corrects
the error.

Also fix the billing demo's smoke test, which runs in CI, to avoid
incorrectly succeeding in the trivial case where no data has yet been
loaded. The smoke test now waits for at least *some* data to load before
asserting that the input and output to the sink have the same number of
records. With this change in place, the test would have reliably failed
in the precense of the above bug.

Also also, rename the CI job from "protobuf kafka streaming-demo" to
"billing demo" to reflect its name elsewhere in the codebase. The name
"billing demo" seems to have won out in the vernacular.