persist: finish Arrangement abstraction merging Unsealed and Trace
This completes the work started in the previous commit of merging the
Unsealed and Trace in-mem structs.

Modeling these two as separate code structures has felt like a mistake
for a while now. This is just one of those things that we got wrong
initially, and the cleanup is overdue.

This makes a persistent collection even more like a differential
dataflow arrangement (arranged collection) than it was previously.
Unsealed is very much analogous to the in-mem buffer that the
arrangement operator keeps data in before the frontier advances past.
And the persisted trace batches are exactly analogous to the in-mem
differential trace batches.

There are two notable differences (besides the obvious durability):
- Because in-mem operations are so much faster than ones on durable
  storage, the act of advancing the frontier and moving data into trace,
  one step in differential, is split into separate steps in persist.
- The differential arrangement keeps the data arranged for efficient
  indexed access (hence the name). Persist also keeps the data arranged
  the same way, but finishing up the plumbing for indexed access is
  still a TODO.

Merging the UnsealedMeta and TraceMeta serializations is left for a
followup PR so we can batch it in with other backward-incompatible
changes.