persist: simplify queue handling in conditional_seal() operator
Before, we kept two queues and would always swing through the whole list
of pending futures when the operator is invoked.

Now, we only keep one `VecDeque` and stop polling the futures on the
first "pending" future.

We only needed the two queues to allow us to know when we were done
checking the in-flight futures, while keeping track of newly added
futures in a separate queue. Otherwise, it could have happened that we
add futures to the queue and never exit the loop.