reuse_topic: Create sink without timing out waiting for message (#9094)
Before this change, we know we've reached the end of the topic by `get_next_message` timing out.  This means it takes a duration of at least `timeout` to create or bootstrap a reuse_topic sink.  Now we look up the offset before reading all the messages and short circuiting when seeing the `hi` watermark.

With this change, the only time we hit the timeout is when the `hi` watermark does not match a message in the kafka topic.

We can see the impact in just the `kafka-exactly-once` testdrive:
```
$ time cargo run --bin testdrive -- test/testdrive/kafka-exactly_once.td
[...]
testdrive completed successfully.

real    0m20.917s
user    0m0.613s
sys     0m0.268s
$ git checkout main
$ time cargo run --bin testdrive -- test/testdrive/kafka-exactly_once.td
[...]
testdrive completed successfully.

real    1m26.748s
user    0m0.625s
sys     0m0.319s
```

### Motivation
https://github.com/MaterializeInc/materialize/issues/8544