expr: auto-infer function output types only in trivial cases
More often than not, functions that have variants (i.e they are not
represented by a zero sized struct) have an output type that depends on
their current variant.

So this patch ensures that when manually implementing the EagerUnaryFunc
trait, which is required for non-ZST functions, a concrete
implementation of their output type is provided.