kafka,source: handle messages received from consumer.poll()
Before, we were assuming that we would never get any messages when
polling the consumer because we split of the partitions into individual
queues. However, when setting an initial partition offset via
`Offset::Offset(o)`, partition assignment and splitting of the queue is
not atomic, meaning we *can* get messages on the main consumer queue.

This refactors `get_next_kafka_message()` to make it more modular and
adds handling for messages that we receive from the main queue via
`consumer.poll()`.