dbt-materialize: use one parameterized spec file
Rather than duplicating the test spec for dbt-materialize to test both
with and without TLS, parameterize it and run it twice.

Also avoid requirements.txt in favor of setup.py to comport with Python
best practices. Notably this avoids installing the `dbt` meta package,
which slims down the Docker image quite a bit.