persist: simplify and fix conditional_seal() operator
Before this change, the operator would let the frontier advance to time
t while it was still waiting for a seal at t to succeed. This is
incorrect, because letting the frontier advance to t signals to
downstream operators that data with a lower timestamp is "ready", while
it is clearly not sealed yet.

Now, we only advance the frontier (by downgrading the capability set)
when a seal is successful.

We'll add a test in a follow-up commit.