repr: introduce a protobuf-backed permanent storage format for Row
We initially intended to use the internal Row encoding for decoding
efficiency, but from a maintenance perspective, it's much easier in
proto to deal with backward and forward compatibility of changes. We can
always change our minds later (with a good bit of glue code, admittedly)
if this turns out to be too much of a bottleneck.