coord,persist: make source persist details more structured/detailed on coordinator
Before, the coordinator would only know the prefix that was used to form
the names of persistent streams that are involved in, say, the
persistent Kafka source.

Now, we keep structured information that is the same for all sources
(`primary_stream` and `timstamp_bindings_stream`) as well as
per-envelope persistence details. This way, the coordinator knows all
persisted streams that are involved in a source.

This also changes what persist information is sent to dataflow workers.
Before, we would send a single persist name. Now, we send the new
`SourcePersistDesc`, which contains the same structured persistence
details that the coordinator knows about. In future commits, we will add
more information to this.

We need to do this to allow the coordinator to:
 - determine a common seal timestamp when restarting
 - negotiate/figure out if a requested as_of matches the compaction
   frontier/since of a persisted source
 - compact persisted streams when necessary

We will do these additional changes in follow-up commits.