coord,persist: teach coord about persisted streams that are involved in sources
Before, the coordinator would only know the "prefix" that was used to
form the names of persistent streams that are involved in, say, the
persistent Kafka source.

Now, we keep a mapping from a _logical name_ to the physical stream name
in the coordinator, for all persisted streams that are involved in a
source.

This also changes what persist information is sent to workers. Before,
we would send a single persist name. Now, we send a new
`SourcePersistDesc`, which contains the mapping of streams. In future
commits, we will add more information to this.

We need to do this to allow the coordinator to:
 - determine a common seal timestamp when restarting
 - negotiate/figure out if a requested as_of matches the compaction
   frontier/since of a persisted source
 - compact persisted streams when necessary

We will do these additional changes in follow-up commits.