persist: correctly set persistent upsert operator name.
Previously, the persisted upsert operator was called "Upsert", which was confusing
because that is the same name as the non-persisted upsert operator, and makes it seem
like that operator is being used in the dataflow. Also, the operator logs a different
operator name to the trace log. This commit changes the operator's dataflow name
to the value sent to trace logs.