persist,table: refactor Tables and SourceDesc to make clear what the persisted_name does
Before, the `persist_name` on SourceDesc was serving two purposes: 1) to
signal that a source actively persists, and 2) to mark a Table as a
persistent table that gets its data by creating a `persisted_source()`
with the given `persist_name`.

Now we clearly separate the two and move the `persist_name` into
`SourceConnector::Local`. It becomes a property of the local source
instead of the descriptor around it.

This will help with future refactoring that change the shape of the
persistence information that we ship with `SourceDesc`.