TAIL: Fix tests
The tests broke because progress and data updates may not be produced
atomically anymore but only in the correct order. FETCH ALL might pick up
data but not yet the associated progress information. This change refactors
the relevant tests to either use a state machine or be more robust to
calling FETCH ALL multiple times.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>