mzcompose: make "wait" step work on modern docker/docker-compose
At least on my machine (see below) a

```
docker-compose ps -q metric-verifier
```

does not print the container id of a container that was manually started
via "run".

With "-a --status running" we seem to get the behaviour that our scripts
expect. I mention in the comment that there is a possible race condition
with this, but I think this was also present in the previous version.

My machine:

```
$ docker --version
Docker version 20.10.11, build dea9396e18
$ docker-compose --version
Docker Compose version 2.2.2
```