persist: fix seal operator to only downgrade after successful seals (#9490)
This commit changes the seal operator to no longer have an error stream, and
instead retry after failed seals, and only downgrade frontiers after a successful
seal.

This commit also adds a test similar to the one for conditional_seal, for
seal. The test verifies that the frontier only successfully advances to t after
a successful seal to time t.