persist: unbreak Kafka upsert sources
Previously, we added kafka metadata into the upsert dataflow. Unfortunately, we only did
so for the non-persisted kafka upsert dataflow, and not for the persisted kafka upsert
dataflow.

This PR aims to fix that by simply folding the metadata into the value which is what
downstream `evaluate` seems to want anyway. The existing upsert implementation likely
does not do so as a performance optimization (no need to create a bunch of new, wider
values if we suspect many of them will be discarded).

This fix is admittedly brittle. A longer term fix would be to unify more of the
code around the persisted codepath, and the nonpersisted codepath. We are leaving
that for followup work.