persist: take persistence since into account when calculating source since
Before, all sources started out with a since frontier of [0] when
created, without looking at the since frontier (aka compaction
frontier) of persisted streams that are involved in a persistent
source.

Now, we look at the since of the primary stream when starting to read
from a source. We only need to look at the primary stream because this
is what carries the actual data of the source that we replay when
starting up.

This was not strictly a bug, because we never compact persistent
sources. We do have to resolve this issue, though, in order to
enable compaction.

Fixes #9655