persist,dataflow: verify expected as_of when replaying streams
Before, it could happen that a dataflow was being rendered at an
expected `as_of` but the persistent streams that were replayed as part
of the dataflow had compacted beyond that. This would manifest in subtle
ways, see for example:
https://github.com/MaterializeInc/materialize/issues/8606.

Now, we thread through the expected `as_of_frontier` and give a
meaningful error message when the expected `as_of` cannot be served.

Note: it's not possible to provoke this new error message from tests. It
will only surface if/when we have a bug in the logic that determines the
as_of frontier (on the coordinator) and forwards it to dataflow
rendering.

Fixes #8608