persist: fix seal() behavior on operator shutdown
Fixes #9419

The seal() operator keeps a queue of pending futures as it issues seal commands. As
those return, it downgrades capabilities corresponding to the times that have now
been sealed. It's careful to do this in order, so as to respect capability downgrade
semantics.

Unfortunately, this had a bug on operator closure where we would drop all capabilities
when we knew the operator was closed (when the input frontier was []), but we did
not drop all of the pending futures. This means that there was a window of time after
we dropped all capabilities when a seal future could invoke the operator again, at
which point we would do an invalid capability downgrade.

This commit fixes the behavior to also drop all pending futures when the operator is closed
and adds a unit test to check regressions.