sql: change context type for plan_{table_function,rows_from}
plan_table_function and plan_rows_from are both called from a relational
context, not a scalar context, so it makes more sense for them to take a
QueryContext rather than an ExprContext. The implementation of
plan_table_function can conjure up the ExprContext when it goes to plan
the arguments to the table function.