sql: constrain scope items to a single name
This better matches how PostgreSQL handles scoping and vastly simplifies
a lot of code that interacts with scopes. The only casualty is a special
case for MySQL that we never should have added in the first place. See
the release note within for details.

Incidentally this also fixes an obscure bug in which whole-row
references would see tables in all outer scopes, rather than just tables
in the first scope to match. I didn't bother adding a release note
because this is so obscure and we don't describe whole-row references
anywhere.