doc/user: make it harder to add release notes to the wrong header
Rework the release notes page so that there is always a dedicated
"Unstable" section. Then old PRs that add release notes will either
conflict or correctly add the release notes to the unstable section.
I'll be responsible for migrating the correct set of release notes to
the dedicated version header when the release ships.