ci: adjust for native macOS M1 builder
I've decided it's easier to run a dedicated macOS M1 builder. This
spares us from having to cross compile Materialize from x86_64 to arm64,
which is always at risk of breaking with every new macOS version or new
C dependency. Importantly, having a macOS M1 builder opens up the
possibility of actually running tests on M1, so we're not just blindly
shipping a binary that might not even be loadable.