*: deoptionalize column names everywhere
As of recently, we no longer permit sources or views to contain unnamed
columns.  Instead, we generate the concrete name "?column?" at the
earliest moment possible, so to the rest of the system it appears that
the user has explicitly named the column "?column?".

So, this commit changes our codebase to use `ColumnName` rather than
`Option<ColumnName>` as the type of a column. This simplifies almost all
code that touches column names, and also avoids a lot of "expect" and
"unwrap" calls whose correctness was hard to verify.