persist: handle trace compactions outside of Indexed
This commit does several things to make trace compactions happen outside of
Indexed.

Indexed loses Maintainer and learns to send MaintenanceReqs after every call
to step (if there are any). When Indexed receives a MaintenanceRes it applies
it via a batched command. Note that in this commit Indexed can spam Maintainer
by repeatedly sending copies of the same request.