persist: only keep one in-flight compaction request at a time
This commit teaches Indexed to only keep one in-flight compaction request
at a time, and to only use a MaintenanceResponses when a stream:

* has an in-flight compaction request
* the request embedded in the MaintenanceResponse matches the in-flight request.

This is important, as using a MaintenanceResponse that does not match the request
that was sent will either:

* not be useful (because the response refers to trace batches that no longer exist)
* be useful, and potentially break the invariant that trace batch compaction levels
  are nonincreasing.

For example, say a stream has trace batches

[a, b, c, d, e, f] and currently has a compaction request out for batches

(c, d). This implies that (a, b) and (b, c) were not compactable because if they
had been, we would have sent a trace compaction request for those instead (the
selection algorithm simply scans pairs left to right looking for two trace batches
not in advance of the compaction frontier that are at the same compaction level).

At this point, if we receive a trace compaction response for (e, f) -> g,
g can be at a new compaction level and if we use that result, turning the list
of trace batches into

[a, b, c, d, g], we can potentially break the invariant that compaction levels are
nonincreasing (if say, a, b, c, d were all at the same compaction level).

This commit leaves a bit of partial work where Error MaintenanceResponses
don't have an embedded request, and as such a mismatched Error response
can still lead Indexed to send duplicate requests. This will be addressed in a
followup commit.