test: add dedicated kafka upsert benchmark
This commit adds a dedicated kafka upsert benchmark using the new mzworkflows
framework but not within the feature-benchmarks framework.

This benchmark allows users to set up a kafka topic and a corresponding upsert
source with byte keys and values and send records to it and observe the time
it takes to ingest the specified number of records.

The benchmark allows for flexibility in the following dimensions:
- key cardinality
- value size
- whether or not Materialize gets restarted after each batch of inserts
- whether or not persistence is enabled
- how many records are sent and how record sending is shaped (e.g. one can
  send many large batches, or send one initial large batch and then many small
  batches of records)

The benchmark currently doesn't support:
- data formats other than bytes
- specifying the number of partitions (currently hardcoded to 4)
- specifying the number of materialize workers