test/chaos: remove
The chaos test in test/chaos has not uncovered many bugs (if any?).
Remove it for now, since it is impeding efforts to switch to
Python-based mzcompose workflows (see #9809). We can re-add a chaos test
in the style of the s3-resumption test when we tackle #7115 in earnest.
We may not even need a dedicated test for this, since our integration
with Antithesis has proven excellent for uncovering these sorts of
issues.

Closes #4558.