kgen: parallelize
Teach kgen to optionally spawn multiple threads, defaulting to the
number of physical CPUs available on the machine.

Thread safety made this surprisingly irritating. This commit refactors
the Avro generator so that the ThreadRng is only ever passed as a
parameter, never stored, as otherwise the Avro generator does not
implement `Send`. It also introduces a rather goofy `Generator` trait
whose only purpose is to make it possible to clone the generator
closures.