[tail] Change where tails communicate having been dropped.
Previously, tails only communicated `Dropped` to the coordinator
after having been fully shut down by Timely. However, since we
were not returning any drop token from the sink, the coordinator
was only cleaning up its inputs, which caused the sinks to naturally
shut down after seeing an empty input frontier. Thus, whether a sink was shutting
down because the dataflow's input was truly exhausted or because
it had been commanded to do so by the coordinator were unidentifiable.

Remedy this by returning a shutdown token for the sink itself,
which will cause it to send a `Dropped` message immediately.