mzcompose: add --preserve-ports flag
The --preserve-ports flag binds all container ports to the same host
ports. This is intended to be used in load tests where we want all
services to be publicly accessible so that they can be scraped by
Prometheus.

As part of this change, disallow binding a container port to a
*different* host port. Either you get a random available port, or you
pass `--preserve-ports` and get a 1:1 mapping. It's then impossible for
two workflows on the same machine to have port conflicts by default. If
you're opting into preserve-ports mode, you wanted to use well-known
ports anyway, so I don't think the loss of flexibility matters.