persist: remove inline metadata in BlobUnsealedBatch
The inlined metadata (either a Range<Seqno> or a Description<u64>) for
Blob{Unsealed|Trace}Batch is only read for validations. It seems like we could
change the API around a little bit to only use the metadata from {Unsealed|Trace}BatchMeta
when validating the corresponding blobs.

Otherwise, I think there's a missing validation somewhere to verify that the
inlined metadata and the separately stored metadata are identical.