persist,sources: allow compaction on persisted sources
The idea behind this is:
 - keep track of what compaction the coordinator allows in `StorageState`
 - received `AllowSourceCompaction` commands update that compaction frontier
 - pass a `Rc<RefCell<>>>` to that frontier to an operator that also
   looks at its input frontier and allows compaction to the combination
   (aka "minimum", or "meet") of those two

An alternative solution would be to invoke `allow_compaction()` upon
receipt of the `AllowSourceCompaction` message. This would, however,
require that we allow async in the command handling code, or to start a
new thread for doing the `allow_compaction` call. Otherwise, we would
block the dataflow worker thread. Starting a thread seems doubly
complicated because we cannot easily hand over the write handle/persist
client to it.

With the chosen solution we only need to update the allowed frontier
centrally and the rendering code that then allows compaction can be
tailored to the rendered source.

Fixes #9508