persist,sources: allow compaction on persisted sources
The idea behind this is:
 - keep track of what compaction the coordinator allows in `RenderState`
 - received `AllowCompaction` commands update that compaction frontier
 - pass a `Rc<RefCell<>>>` to that frontier to an operator that also
   looks at its input frontier and allows compaction to the combination
   (aka "minimum") of those two

An alternative solution would be to invoke `allow_compaction()` upon
receipt of the `AllowCompaction` message. This would, however, require
to keep the involved persistence write handles in `RenderState`, and the
number and type of those handles will likely be different for different
combinations of physical source (think Kafka etc.) and envelopes.

With the chosen solution we only need to update the allowed frontier
centrally and the rendering code that then allows compaction can be
tailored to the rendered source.

Fixes #9508