pg-cdc: check that the publication exists
Using the query right after the one added in this commit could still
lead to a valid result if the publication name doesn't exist. It simply
appeared as a publication that has no tables attached to it.

This query checks for the existance specifically and errors out if the
publication is missing.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>