sql: use ExprContext name as cast op
plan_op previously took an `op: &str` parameter which was used in error
messages to provide context about which part of the SQL query included
the invalid cast. This is exactly what the ExprContext::name field is
for, and in fact in several cases `&ecx.name` was already supplied as
the `op`.