coord: move system catalog objects to their own timeline
Change the transaction guarantees for system catalog objects to remove
snapshot isolation and instead always return the latest state. This
allows them to skip all transaction machinery and always be runnable
in any transaction or timedomain. This unblocks a PGJDBC feature which
fetches type information for unknown types needed by some tools.

As a result, catalog tables are no longer joinable with user tables. We
believe this is an acceptable tradeoff because we gain the ability to
always query the catalog regardless of which timeline the transaction
is in. Even with some solution that, say, always included catalog tables
in all realtime timeline transactions, users with non-realtime timeline
data would not be able to use the catalog should their driver need that
(like PGJDBC).

Fixes #9375
Fixes #9374