sinks: only borrow the frontier refcell temporarily
The code was previously unconditionally borrowing the RefCell wrapping
the frontiers causing the async operator be unable to update them.

The orignal type for the frontiers was
`Rc<RefCell<Vec<Antichain<Timestamp>>>>` which made it difficult to
split the frontiers of each input without permanently borrowing the
RefCell. This inconvenience was probably one of the reasons for the
preexisitng pattern.

This PR changes the type of frontiers to be
`Vec<Rc<RefCell<Antichain<Timestamp>>>>` so that the frontiers of each
input can be picked apart easily without borrowing the individual
`RefCell`s and all uses of the frontier in the async operator have their
respective `.borrow()` call right at the use site.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>