testdrive: merge workflows into one ergonomic pile
Merge all the testdrive workflows into a single `testdrive` workflow
that takes command-line flags. This makes it much more ergonomic to run
`testdrive` in various configurations from the command line.

I didn't see any downside to merging the Redpanda support back into the
main testdrive composition, because Redpanda is only launched when the
`--redpanda` option is supplied. This means that folks on M1 Macs can
use testdrive almost as easily as folks on Linux:

    ./mzcompose --dev run testdrive --redpanda foo.td

I updated the documentation accordingly, and dropped in a few notes
about how unusably slow the Confluent Platform is on ARM machines.

Finally, I got rid of the `shell_eval` parameter to Testdrive, because
it was well beyond my capacity to understand. Instead, `testdrive` the
binary has learned to perform its own glob expansion, which covers 95%
of cases. The last 5% is the Redpanda nightly job, which needs a custom
workflow to handle filtering out the kafka-offset-times.td test file
which is not yet supported. This custom logic all lives in Python now,
though, so I think it's much easier to understand.