python/spawn: mandate keyword arguments in `runv`/`capture`
When developing, it's easy to accidentally type

    spawn.runv("command", "arg")

instead of:

    spawn.runv(["command", "arg"])

Previously the former would fail with a confusing message about how
"arg" is not of expected type "Path" because "arg" was interpreted as
the value of the `cwd` parameter. By mandating that all options but the
args be passed by keyword (which 100% of checked-in callers already do)
we get a more understandable error message.