interchange,testdrive: use prost-reflect instead of rust-protobuf
Split out from #10079. See that PR for rationale on the migration to
prost.

This PR swaps dynamic message encoding/decoding from rust-protobuf to
prost-reflect. The APIs are quite similar, so the conversion is
straightforward.

Unfortunately, prost-reflect does not make it easy to perform the strict
message equivalence checking we were performing previously, where once
we saw a message with a given message descriptor, we'd never accept a
message with a different descriptor. Rather than tackling that problem,
I figured I'd instead allow for schema evolution (#9598) in the same way
that the official Confluent client libraries do.

The idea is to simply attempt to decode the protobuf without checking
the schema ID or message descriptor. If you're evolving your Protobufs
backwards compatibly, this works great, and we follow the normal
Protobuf rules for handling e.g. integer promotion/truncation. If not...
you'll get an error, if you're lucky, or garbage data otherwise. In most
cases, folks have their schema registries configured to prevent them
from evolving their schemas incompatibily, so I think this is the right
decision.

I've added a test for this new behavior and removed the test for the old
behavior.

This commit willfix the second work item in #9598.