test/proxy: support redpanda
Support a `--redpanda` option in test/proxy just like in test/testdrive.
This is mostly for the convenience of folks on M1 Macs who can't easily
run Confluent Platform in Docker.

While I'm in here, use the new `c.override` feature to avoid creating
separate `Materialized` services, and imbue the workflow with the same
`--aws-region` option that test/testdrive has.

There's a bit of duplication with test/testdrive now, but I want to
avoid introducing any helpers until we've got a few more examples of
this duplication so that we can get the abstraction boundary right.