Improve error message specificity for AWS initialization errors
Previously the purification step would fail with a message that contained just
"Unable to build connector." Which doesn't explain why we're trying to build a
connector, what kind of connector it is, or provide resolution steps.

With this, we know that it's failing to validate, it will say something like
"Access Denied" if it's not possible to access STS, etc.