Switch from rust-protobuf to protobuf-native for proto compilation
See #10079 for the justification for switching away from rust-protobuf.

This commit is the final piece of the puzzle. It switches sql and
coord's protobuf compilation from rust-protobuf to protobuf-native.  It
also inlines the logic in those crates, since interchange no longer
needs to compile Protobuf schemas dynamically.

protobuf-native is a thin layer on top of the official C++ libprotobuf
library. It is exactly equivalent to invoking `protoc`, except that the
input files don't need to be written to a temporary file on disk, which
makes for a much nicer API.