dataflow,persist: filter and retract timestamp bindings that are beyond the common seal ts
Before, we were keeping bindings beyond the common seal  timestamp, but
it would lead to problems where a restarting source would advance the
seal frontier and bindings that were not valid suddenly become valid
after a second restart.

Now, we make sure to retract those bindings that are beyond the common
seal timestamp, just as we do for persisted data/updates that are beyond
the common seal timestamp.

This also fixes the consolidation of timestamp bindings on restore:

Before, we were not taking the `diff` field of restored bindings into
account when determining the starting offsets. This could lead to the
following case:

uppser_seal_ts: 5
binding updates:
((offset1, ts1), 3, 1)
((offset2, ts2), 4, 1)
((offset2, ts2), 4, -1)

Here, the correct offset to start reading from is `offset` bout we would
mistakenly use `offset2`.

Now, we collect all bindings updates that are not beyond the common
uppser seal timestamp, consolidate them, and then determine the starting
offset from what remains, that is, we only use bindings whose `diff` is
`1`.

Fixes #9687