sink: avoid needlessly capturing self lifetimes
All the producer methods internally clone the inner producer and produce
a future that doesn't capture &self. This is something that is not
captured by the `async fn` which leads to errors down the line when
trying to use `bail_err!()`.

Using the `-> impl Future<..>` syntax we're able to express that the
returned future does not capture the input lifetime from &self.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>