Fix timestamp logic around metadata availability
After a refactor, this code ended up inserting metadata such that it
would turn valid only _after_ the "retain_for" had passed. This is
usually ok in the normal case, but means that for the first n scrape
intervals, there is no metrics metadata available (because it only
starts being valid after the batch for which it got inserted gets
expired).

Instead of this wrong behavior, insert the metadata update (and
expiry) to be valid as of now, and remove the confusing argument.