testdrive: Allow passing in variables at the command line
For an upcoming mzcompose test I need to dynamically generate AWS resources and
pass them into test drive.

It is currently possible to expose them to testdrive tests by explicitly
setting env vars, but from inside of mzcompose.py it's less clear that we're
passing variables that need to be used by the td file and not just setting
things that affect the way testdrive behaves.